Bedre typesikring på henting av data Story
