+++ packages/v2/backend/package.json
@@ -16,6 +16,6 @@
   },
   "dependencies": {
     "@navikt/k9-klage-typescript-client": "1.0.20240513162434-997f3da",
-    "@navikt/k9-sak-typescript-client": "1.0.20240916120432"
+    "@navikt/k9-sak-typescript-client": "1.0.20241021094353"
   }
 }
