+++ packages/v2/gui/src/prosess/ung-beregning/SatserData.ts
@@ -1,0 +1,8 @@
+export interface SatserData {
+  fom: string;
+  tom: string;
+  dagsats: number;
+  grunnbeløpFaktor: number;
+  grunnbeløp: number;
+  satsType: 'LAV' | 'HØY';
+}
