+++ packages/v2/gui/src/prosess/ung-beregning/UngBeregningBackendApiType.ts
@@ -1,0 +1,5 @@
+import type { SatserData } from './SatserData';
+
+export type UngBeregningBackendApiType = {
+  getSatser(behandlingUuid: string): Promise<SatserData[]>;
+};
